+++ nodejs/packagescanner.go
@@ -45,7 +45,7 @@ type Scanner struct{}
 func (*Scanner) Name() string { return "nodejs" }
 
 // Version implements scanner.VersionedScanner.
-func (*Scanner) Version() string { return "3" }
+func (*Scanner) Version() string { return "4" }
 
 // Kind implements scanner.VersionedScanner.
 func (*Scanner) Kind() string { return "package" }
