+++ ruby/ecosystem.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/quay/claircore/indexer"
+	"github.com/quay/claircore/language"
 )
 
 var scanners = []indexer.PackageScanner{&Scanner{}}
@@ -14,6 +15,6 @@ func NewEcosystem(_ context.Context) *indexer.Ecosystem {
 		PackageScanners:      func(_ context.Context) ([]indexer.PackageScanner, error) { return scanners, nil },
 		DistributionScanners: func(_ context.Context) ([]indexer.DistributionScanner, error) { return nil, nil },
 		RepositoryScanners:   func(_ context.Context) ([]indexer.RepositoryScanner, error) { return nil, nil },
-		Coalescer:            NewCoalescer,
+		Coalescer:            language.NewCoalescer,
 	}
 }
